Move swapperd.ts and sleep.ts to common folder
