+++ src/renderer/components/BalanceItem.tsx
@@ -4,7 +4,7 @@ import BigNumber from "bignumber.js";
 import { CopyToClipboard } from "react-copy-to-clipboard";
 
 import { getLogo } from "@/lib/logos";
-import { IPartialWithdrawRequest, Token } from "@/lib/swapperd";
+import { IPartialWithdrawRequest, Token } from "common/swapperd";
 
 interface IBalanceItemProps {
     token: Token;
